+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Export the generated palettes from the palette dialog as a GIMP palette file
 - A `.doap` file to describe the project 
+- LMS color space can now be shown
 
 ### Changed
 - Fixed a bug where the illuminants where not fully shown